+++ src/call.ts
@@ -25,7 +25,7 @@ import {
   ResultTuple,
   SimpleCallbackFunction,
 } from './apitypes';
-import {GoogleError} from './googleError';
+import {GoogleError, GoogleErrorDecoder} from './googleError';
 
 export class OngoingCall {
   callback: APICallback;
@@ -78,11 +78,18 @@ export class OngoingCall {
     if (this.completed) {
       return;
     }
-    // eslint-disable-next-line
-    const canceller = func(argument, (...args: any[]) => {
-      this.completed = true;
-      setImmediate(this.callback!, ...args);
-    });
+    const canceller = func(
+      argument,
+      (
+        err: GoogleError | null,
+        response?: ResponseType,
+        next?: NextPageRequestType,
+        rawResponse?: RawResponseType
+      ) => {
+        this.completed = true;
+        setImmediate(this.callback!, err, response, next, rawResponse);
+      }
+    );
     this.cancelFunc = () => canceller.cancel();
   }
 }
@@ -111,7 +118,14 @@ export class OngoingCallPromise extends OngoingCall {
       rawResponse?: RawResponseType
     ) => {
       if (err) {
-        rejectCallback(err);
+        const decoder = new GoogleErrorDecoder();
+        try {
+          const decodedErr = decoder.decodeMetadata(err);
+          rejectCallback(decodedErr);
+        } catch (decodeErr) {
+          // Ignore the decoder error now, and return back original error.
+          rejectCallback(err);
+        }
       } else if (response !== undefined) {
         resolveCallback([response, next || null, rawResponse || null]);
       } else {